Moved fetch-wrapper.tsx outside of _lib and deleted _lib.
